Change in mcr API, now you can use double[][][] for cost function
git-svn-id: svn://svn.code.sf.net/p/choco/code/trunk@1388 5779b9a9-2447-0410-887e-bb98d57c6de4

+++ choco-solver/choco-cp/src/main/java/choco/cp/model/managers/constraints/global/MultiCostRegularManager.java
@@ -61,9 +61,9 @@ public class MultiCostRegularManager extends IntConstraintManager
                 System.arraycopy(all, 0, vs, 0, vs.length);
                 System.arraycopy(all, vs.length, z, 0, all.length - vs.length);
 
-                if (param[2] instanceof int[][][])
+                if (param[2] instanceof double[][][])
                 {
-                    int[][][] csts = (int[][][]) param[2];
+                    double[][][] csts = (double[][][]) param[2];
                     return new FastMultiCostRegular(vs,z,pi,csts);
                 }
               /*  else if (param[2] instanceof int[][][][])
